refactor: collect identity endpoints
